fix(mission) mission api fix

+++ src/main/java/io/github/wotjd243/findbyhint/MissionInventory/domain/MissionInventoryInfo.java
@@ -12,7 +12,7 @@ import java.math.BigInteger;
 public class MissionInventoryInfo {
 
     @Column(nullable = false)
-    private BigInteger missionId;
+    private Long missionId;
 
 //    @Embedded
 //    private MissionQnA missionQnA;
@@ -26,14 +26,14 @@ public class MissionInventoryInfo {
     private MissionInventoryInfo() {
     }
 
-    public MissionInventoryInfo(BigInteger missionId, String question, String answer) {
+    public MissionInventoryInfo(Long missionId, String question, String answer) {
         this.missionId = missionId;
         this.question = question;
         this.answer = answer;
         this.status = MissionSuccessStatus.FAIL;
     }
 
-    public static MissionInventoryInfo valueOf(BigInteger missionId, String question, String answer) {
+    public static MissionInventoryInfo valueOf(Long missionId, String question, String answer) {
         return new MissionInventoryInfo(missionId, question, answer);
     }
